refactor - move InstallSdkAction.java into flutter/module
