+++ src/io/flutter/module/FlutterGeneratorPeer.java
@@ -22,7 +22,6 @@ import com.intellij.ui.components.labels.LinkLabel;
 import com.intellij.util.ReflectionUtil;
 import com.intellij.xml.util.XmlStringUtil;
 import io.flutter.FlutterBundle;
-import io.flutter.actions.InstallSdkAction;
 import io.flutter.sdk.FlutterSdkUtil;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
